change the stock last price model to float

+++ api/models.py
@@ -2,4 +2,4 @@ from django.db import models
 
 class Stock(models.Model):
     ticker = models.CharField(max_length=30,unique=True)
-    last_price = models.IntegerField(default=0)
+    last_price = models.FloatField(default=0)
